style: 예제 테스트 실행 오류로 인한 테스트 재배치
