+++ src/test/java/racingcar/ApplicationTest.java
@@ -1,6 +1,8 @@
 package racingcar;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
+import java.util.Arrays;
+import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
@@ -9,7 +11,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class ApplicationTest extends NsTest {
-    private static final int MOVING_FORWARD = 4;
+    private static final int MOVING_FORWARD = 4;실
     private static final int STOP = 3;
 
     @Test
@@ -21,6 +23,23 @@ class ApplicationTest extends NsTest {
             },
             MOVING_FORWARD, STOP
         );
+    }
+
+    @Test
+    void split_입력테스트1() {
+        String input = ",,,,,";
+        List<String> result = Arrays.asList(input.split(","));
+
+        assertThat(result).isEmpty();
+    }
+
+    @Test
+    void split_입력테스트2() {
+        String input = "hwang,seong,heon,,,";
+        List<String> result = Arrays.asList(input.split(","));
+
+        assertThat(result.size()).isEqualTo(3);
+        assertThat(result).containsExactly("hwang","seong","heon");
     }
 
     @Test
